CoffeeShop, Class 새로운 swift파일로 이전 Coffee enum 중복 변수명 제거
